requirements.txt
# 库街区登录项目依赖管理
# 主依赖文件 - 包含所有功能所需依赖

# 基础依赖
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0
requests>=2.32.0
loguru>=0.7.0
pycryptodome>=3.20.0

# 图像处理和验证码识别
opencv-python>=4.8.0
Pillow>=10.0.0
ddddocr>=1.5.0

# 机器学习和AI（用于图标识别）
torch>=2.0.0
torchvision>=0.15.0
open_clip_torch>=2.20.0
sentence-transformers>=2.2.0
numpy>=1.24.0

# 工具库
pydash>=7.0.0
urllib3>=2.0.0
python-dotenv>=1.0.0

# 开发依赖（可选）
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0